[Bugfix] Fix int32 address overflow in DeepGEMM ep_gather Triton kernel

+++ vllm/model_executor/layers/fused_moe/deep_gemm_utils.py
@@ -389,9 +389,12 @@ def _fwd_kernel_ep_gather(
                 expert_id = apply_expert_map(expert_id, expert_map)
 
             if expert_id >= 0:
+                # int64: source rows live in the expert-sorted (padded) buffer,
+                # whose row offset can exceed int32 range for large-batch
+                # EP/DP shapes (source_token_index * stride0 > 2**31 - 1).
                 source_token_index = tl.load(
                     input_index + cur_token * input_index_stride0 + topk_index
-                )
+                ).to(tl.int64)
                 acc_weight = tl.load(
                     recv_topk_weight + cur_token * recv_topk_weight_stride0 + topk_index
                 )
@@ -405,7 +408,7 @@ def _fwd_kernel_ep_gather(
 
         tl.store(
             output_tensor
-            + cur_token * output_tensor_stride0
+            + cur_token.to(tl.int64) * output_tensor_stride0
             + cur_block * BLOCK_D
             + off_d,
             accumulator.to(output_tensor.dtype.element_ty),
